Move get_bundled_aggregations to relations handler.
