+++ changelog.d/12237.misc
@@ -1,0 +1,1 @@
+Refactor the relations endpoints to add a `RelationsHandler`.
